Move emit from extract to astToIR convert
This makes us ready to delete the extract.ts Converter class
